Change directory of Bybit stores
