--- pybotters/models/experimental/__init__.py
@@ -1,7 +1,0 @@
-from typing import Tuple
-from .bybit import BybitInverseDataStore, BybitUSDTDataStore
-
-__all__: Tuple[str, ...] = (
-    'BybitInverseDataStore',
-    'BybitUSDTDataStore',
-)
